requirements.txt
#langchain core
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-classic
langchain-chroma
pydantic
pydantic[email]
pypdf


#retrievers
wikipedia
selenium
beautifulsoup4
unstructured


#openai-integration
langchain-openai
openai


#anthropic integration
langchain-anthropic


#google gemini (PaLM) integration
langchain-google-genai>=2.0.0
google-generativeai
youtube-transcript-api


#hugging face integration
langchain-huggingface
transformers
huggingface-hub
sentence-transformers


#Download purpose
hf_xet
accelerate
bitsandbytes


#Web interface
streamlit
grandalf


#environment variable management
python-dotenv


#Vector database
chromadb
tiktoken
langchain-chroma


#machine learning utilities
numpy
pandas
scikit-learn
scipy